// src/app/model/entity/Cliente.ts
export interface Cliente {

    id: number;
    nome: string;
    cpf: string;
    email: string;
    dataNascimento: string;
    telefone: string;
    telefone2: string;
    cep: string;
    logradouro: string;
    numero: string;
    cidade: string;
    estado: string;
    pais: string;
    situacao: string;
}